Fix platform

+++ src/drawing.rs
@@ -22,7 +22,7 @@ impl Vertex {
 #[derive(Default)]
 pub struct DrawList {
 	pub vertices: Vec<Vertex>,
-	pub indices:  Vec<u16   >,
+	pub indices:  Vec<u32>,
 }
 
 impl DrawList {
@@ -32,7 +32,7 @@ impl DrawList {
 	}
 
 	#[inline]
-	fn idx_tri(&mut self, a: u16, b: u16, c: u16) {
+	fn idx_tri(&mut self, a: u32, b: u32, c: u32) {
 		self.indices.push(a);
 		self.indices.push(b);
 		self.indices.push(c);
@@ -40,7 +40,7 @@ impl DrawList {
 
 	/// adds a rectangle\
 	pub fn add_rect(&mut self, rect: Rect, color: u32) {
-		let offset = self.vertices.len() as u16;
+		let offset = self.vertices.len() as u32;
 		self.vertices.push(Vertex::new(rect.min, color)); // topleft
 		self.vertices.push(Vertex::new(vec2(rect.max.x, rect.min.y), color)); // topright
 		self.vertices.push(Vertex::new(vec2(rect.min.x, rect.max.y), color)); // bottomleft
@@ -51,7 +51,7 @@ impl DrawList {
 
 	/// adds a triangle on the bottom right, usuaully used for the resizing thing
 	pub fn add_bottom_right_triangle(&mut self, rect: Rect, color: u32) {
-		let offset = self.vertices.len() as u16;
+		let offset = self.vertices.len() as u32;
 		self.vertices.push(Vertex::new(vec2(rect.max.x, rect.min.y), color)); // topright
 		self.vertices.push(Vertex::new(vec2(rect.min.x, rect.max.y), color)); // bottomleft
 		self.vertices.push(Vertex::new(rect.max, color)); // bottomright
